refactor(app): extract community hierarchy
